refactor: create an account subdirectory in the model directory
